Modernize to C++17 with compiler hardening
Add OpenSSF compiler hardening flags (checked for support at configure
time) and upgrade the codebase from C++11 to idiomatic C++17.

Hardening: -Wconversion, -Wsign-conversion, -Wformat=2,
-Wimplicit-fallthrough, -Werror=format-security, -fcf-protection=full,
-fstack-clash-protection, -fstack-protector-strong, -fPIE/-pie,
-ftrivial-auto-var-init=zero, full RELRO, and _FORTIFY_SOURCE=3 for
optimized builds.  -Werror is Debug-only.

C++17 structural changes:
- Default member initializers on all structs; eliminate memset-over-
  struct UB (machine_t contained a std::string via uart_t)
- std::optional<int> for v20_t::seg_override (replaces -1 sentinel)
- enum class for FdcPhase, UartBackend, I8251State, CentBackend
- std::unique_ptr<FILE, FClose> (FilePtr) for fdc_t::disk and
  machine_t::printer — automatic cleanup via RAII
- std::array for all fixed-size arrays in v20_t, fdc_t, machine_t,
  rtc_t (breakpoints, trace buffer, cmd/res, banks, keyboard rows)
- constexpr replacing #define for size/clock constants
- [[nodiscard]] on error-returning public API functions
- strcpy replaced with snprintf in the disassembler

+++ src/dbg_dis.cpp
@@ -99,7 +99,7 @@ void DbgDisPanel::disassemble_region()
 		char hex[32] = "";
 		int hpos = 0;
 		for (int j = 0; j < len && j < 6; j++)
-			hpos += snprintf(hex + hpos, sizeof(hex) - hpos, "%02X ", code[j]);
+			hpos += snprintf(hex + hpos, sizeof(hex) - (size_t)hpos, "%02X ", code[j]);
 
 		char marker = (m_seg == m_cpu->cs && off == m_cpu->ip) ? '>' : ' ';
 		snprintf(line, sizeof(line), "%c%04X:%04X  %-18s %s\n",
